requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
cycler==0.10.0
    # via matplotlib
deap==1.3.1
    # via -r requirements.in
joblib==1.0.1
    # via scikit-learn
kiwisolver==1.3.2
    # via matplotlib
matplotlib==3.4.3
    # via qdpy
numpy==1.21.2
    # via
    #   -r requirements.in
    #   deap
    #   matplotlib
    #   pandas
    #   qdpy
    #   scikit-learn
    #   scipy
pandas==1.3.3
    # via
    #   -r requirements.in
    #   qdpy
pillow==8.3.2
    # via matplotlib
psutil==5.8.0
    # via qdpy
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2021.1
    # via pandas
pyyaml==5.4.1
    # via qdpy
qdpy==0.1.2.1
    # via -r requirements.in
scikit-learn==0.24.2
    # via sklearn
scipy==1.7.1
    # via
    #   qdpy
    #   scikit-learn
six==1.16.0
    # via
    #   cycler
    #   python-dateutil
sklearn==0.0
    # via qdpy
threadpoolctl==2.2.0
    # via scikit-learn
typing-extensions==3.10.0.2
    # via qdpy
gym==0.20.0
git+https://github.com/benelot/pybullet-gym	
